Update ZSH prompt
This is fairly closely based on an example from the Arch wiki. The major
change is that it includes the date stamp.
https://wiki.archlinux.org/index.php/Zsh#Example

The motivation for this change was that the previous prompt would lose
lines that had no trailing newline. This prompt (as well as most others)
does not have that issue.