Add yq disambiguation wrapper
I have issues remembering that there are two distinct `yq`s that behave
slightly differently, in incompatible ways. This is a reminder. In the
future, I'd like to automatically detect and switch between them; for
now, I just print a warning every time they're used. (Another
potential enhancement would be to make this a shell function that's only
used in interactive scripts.)